# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NuweFarma API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar todos los roles",
                "operationId": "5c93e433a52fc61578863a28b96cc3e2",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "permiso_id": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "nullable": true
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "536fc2c14188dbfbd82a6bf634f31fe1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "estado"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Rol con acceso total",
                                        "nullable": true
                                    },
                                    "permiso_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Ver detalles de un rol",
                "operationId": "b2cb2d6ae0f5ddcc411843aa788476ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del rol"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Editar un rol",
                "operationId": "60b1b72ad61d37b2d670d01cac35d7ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "permiso_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "nullable": true
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "activo",
                                            "inactivo"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "operationId": "c13c8b08ddc181f9c9ecd5935be3b759",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            }
        },
        "/api/roles/{id}/toggle-estado": {
            "patch": {
                "tags": [
                    "Roles"
                ],
                "summary": "Activar o desactivar un rol",
                "operationId": "6ca7de0c377e0ba74680f7e07d5054f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del rol actualizado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}